\name{timeseries}
\alias{timeseries}
\docType{data}
\title{
Time series example
}
\description{
blabla
}
\usage{data(timeseries)}
\format{
  A data frame with 79 observations on the following 397 variables.
  \describe{
\code{timeseries} is a data frame with many variables :\cr
\code{name} is a factor with levels \code{Assandh, India} \code{Gharaunda} \code{Goganda,India} \code{Khansama, Bangladesh} \code{Phulbari, Bangladesh} \code{Sarada,India}.\cr
All the other fields are the time series with the name following the notation :
XYYYYMMP
YYYY the year (from 1999 to 2009
MM the month (between 01 and 12)
P the period (between 0 and 2).  
  }
}
\details{
no details yet.
}
\source{
VITO
}
\examples{
data(timeseries)
## maybe str(timeseries) ; plot(timeseries) ...
}
\keyword{datasets}
