\name{mapdiff}
\alias{mapdiff}
\title{Maps differences}
\description{Compute differences of 2 maps
}
\usage{
mapdiff(mapin, mapref, mapsd = FALSE, fileout = "diff.tif")
}
\arguments{
  \item{mapin}{map 1}
  \item{mapref}{map 2}
  \item{mapsd}{to take into account the standard deviation : not available yet}
  \item{fileout}{where the result is saved}
}
\details{explanation of the formula}
\value{
Returns the resulting map
}
\author{Romain Frelat and Bruno Gerard}
\keyword{spatial}
