\name{timeseries}
\alias{timeseries}
\docType{data}
\title{
Time series example
}
\description{
blabla
}
\usage{data(timeseries)}
\format{
  A data frame with 79 observations on the following 397 variables.
  \describe{
    \item{\code{name}}{a factor with levels \code{Assandh, India} \code{Gharaunda} \code{Goganda,India} \code{Khansama, Bangladesh} \code{Phulbari, Bangladesh} \code{Sarada,India}}
    \item{\code{X1999010}}{a numeric vector}
    \item{\code{X1999011}}{a numeric vector}
    \item{\code{X1999012}}{a numeric vector}
    \item{\code{X1999020}}{a numeric vector}
    \item{\code{X1999021}}{a numeric vector}
    \item{\code{X1999022}}{a numeric vector}
    \item{\code{X1999030}}{a numeric vector}
    \item{\code{X1999031}}{a numeric vector}
    \item{\code{X1999032}}{a numeric vector}
    \item{\code{X1999040}}{a numeric vector}
    \item{\code{X1999041}}{a numeric vector}
    \item{\code{X1999042}}{a numeric vector}
    \item{\code{X1999050}}{a numeric vector}
    \item{\code{X1999051}}{a numeric vector}
    \item{\code{X1999052}}{a numeric vector}
    \item{\code{X1999060}}{a numeric vector}
    \item{\code{X1999061}}{a numeric vector}
    \item{\code{X1999062}}{a numeric vector}
    \item{\code{X1999070}}{a numeric vector}
    \item{\code{X1999071}}{a numeric vector}
    \item{\code{X1999072}}{a numeric vector}
    \item{\code{X1999080}}{a numeric vector}
    \item{\code{X1999081}}{a numeric vector}
    \item{\code{X1999082}}{a numeric vector}
    \item{\code{X1999090}}{a numeric vector}
    \item{\code{X1999091}}{a numeric vector}
    \item{\code{X1999092}}{a numeric vector}
    \item{\code{X1999100}}{a numeric vector}
    \item{\code{X1999101}}{a numeric vector}
    \item{\code{X1999102}}{a numeric vector}
    \item{\code{X1999110}}{a numeric vector}
    \item{\code{X1999111}}{a numeric vector}
    \item{\code{X1999112}}{a numeric vector}
    \item{\code{X1999120}}{a numeric vector}
    \item{\code{X1999121}}{a numeric vector}
    \item{\code{X1999122}}{a numeric vector}
    \item{\code{X2000010}}{a numeric vector}
    \item{\code{X2000011}}{a numeric vector}
    \item{\code{X2000012}}{a numeric vector}
    \item{\code{X2000020}}{a numeric vector}
    \item{\code{X2000021}}{a numeric vector}
    \item{\code{X2000022}}{a numeric vector}
    \item{\code{X2000030}}{a numeric vector}
    \item{\code{X2000031}}{a numeric vector}
    \item{\code{X2000032}}{a numeric vector}
    \item{\code{X2000040}}{a numeric vector}
    \item{\code{X2000041}}{a numeric vector}
    \item{\code{X2000042}}{a numeric vector}
    \item{\code{X2000050}}{a numeric vector}
    \item{\code{X2000051}}{a numeric vector}
    \item{\code{X2000052}}{a numeric vector}
    \item{\code{X2000060}}{a numeric vector}
    \item{\code{X2000061}}{a numeric vector}
    \item{\code{X2000062}}{a numeric vector}
    \item{\code{X2000070}}{a numeric vector}
    \item{\code{X2000071}}{a numeric vector}
    \item{\code{X2000072}}{a numeric vector}
    \item{\code{X2000080}}{a numeric vector}
    \item{\code{X2000081}}{a numeric vector}
    \item{\code{X2000082}}{a numeric vector}
    \item{\code{X2000090}}{a numeric vector}
    \item{\code{X2000091}}{a numeric vector}
    \item{\code{X2000092}}{a numeric vector}
    \item{\code{X2000100}}{a numeric vector}
    \item{\code{X2000101}}{a numeric vector}
    \item{\code{X2000102}}{a numeric vector}
    \item{\code{X2000110}}{a numeric vector}
    \item{\code{X2000111}}{a numeric vector}
    \item{\code{X2000112}}{a numeric vector}
    \item{\code{X2000120}}{a numeric vector}
    \item{\code{X2000121}}{a numeric vector}
    \item{\code{X2000122}}{a numeric vector}
    \item{\code{X2001010}}{a numeric vector}
    \item{\code{X2001011}}{a numeric vector}
    \item{\code{X2001012}}{a numeric vector}
    \item{\code{X2001020}}{a numeric vector}
    \item{\code{X2001021}}{a numeric vector}
    \item{\code{X2001022}}{a numeric vector}
    \item{\code{X2001030}}{a numeric vector}
    \item{\code{X2001031}}{a numeric vector}
    \item{\code{X2001032}}{a numeric vector}
    \item{\code{X2001040}}{a numeric vector}
    \item{\code{X2001041}}{a numeric vector}
    \item{\code{X2001042}}{a numeric vector}
    \item{\code{X2001050}}{a numeric vector}
    \item{\code{X2001051}}{a numeric vector}
    \item{\code{X2001052}}{a numeric vector}
    \item{\code{X2001060}}{a numeric vector}
    \item{\code{X2001061}}{a numeric vector}
    \item{\code{X2001062}}{a numeric vector}
    \item{\code{X2001070}}{a numeric vector}
    \item{\code{X2001071}}{a numeric vector}
    \item{\code{X2001072}}{a numeric vector}
    \item{\code{X2001080}}{a numeric vector}
    \item{\code{X2001081}}{a numeric vector}
    \item{\code{X2001082}}{a numeric vector}
    \item{\code{X2001090}}{a numeric vector}
    \item{\code{X2001091}}{a numeric vector}
    \item{\code{X2001092}}{a numeric vector}
    \item{\code{X2001100}}{a numeric vector}
    \item{\code{X2001101}}{a numeric vector}
    \item{\code{X2001102}}{a numeric vector}
    \item{\code{X2001110}}{a numeric vector}
    \item{\code{X2001111}}{a numeric vector}
    \item{\code{X2001112}}{a numeric vector}
    \item{\code{X2001120}}{a numeric vector}
    \item{\code{X2001121}}{a numeric vector}
    \item{\code{X2001122}}{a numeric vector}
    \item{\code{X2002010}}{a numeric vector}
    \item{\code{X2002011}}{a numeric vector}
    \item{\code{X2002012}}{a numeric vector}
    \item{\code{X2002020}}{a numeric vector}
    \item{\code{X2002021}}{a numeric vector}
    \item{\code{X2002022}}{a numeric vector}
    \item{\code{X2002030}}{a numeric vector}
    \item{\code{X2002031}}{a numeric vector}
    \item{\code{X2002032}}{a numeric vector}
    \item{\code{X2002040}}{a numeric vector}
    \item{\code{X2002041}}{a numeric vector}
    \item{\code{X2002042}}{a numeric vector}
    \item{\code{X2002050}}{a numeric vector}
    \item{\code{X2002051}}{a numeric vector}
    \item{\code{X2002052}}{a numeric vector}
    \item{\code{X2002060}}{a numeric vector}
    \item{\code{X2002061}}{a numeric vector}
    \item{\code{X2002062}}{a numeric vector}
    \item{\code{X2002070}}{a numeric vector}
    \item{\code{X2002071}}{a numeric vector}
    \item{\code{X2002072}}{a numeric vector}
    \item{\code{X2002080}}{a numeric vector}
    \item{\code{X2002081}}{a numeric vector}
    \item{\code{X2002082}}{a numeric vector}
    \item{\code{X2002090}}{a numeric vector}
    \item{\code{X2002091}}{a numeric vector}
    \item{\code{X2002092}}{a numeric vector}
    \item{\code{X2002100}}{a numeric vector}
    \item{\code{X2002101}}{a numeric vector}
    \item{\code{X2002102}}{a numeric vector}
    \item{\code{X2002110}}{a numeric vector}
    \item{\code{X2002111}}{a numeric vector}
    \item{\code{X2002112}}{a numeric vector}
    \item{\code{X2002120}}{a numeric vector}
    \item{\code{X2002121}}{a numeric vector}
    \item{\code{X2002122}}{a numeric vector}
    \item{\code{X2003010}}{a numeric vector}
    \item{\code{X2003011}}{a numeric vector}
    \item{\code{X2003012}}{a numeric vector}
    \item{\code{X2003020}}{a numeric vector}
    \item{\code{X2003021}}{a numeric vector}
    \item{\code{X2003022}}{a numeric vector}
    \item{\code{X2003030}}{a numeric vector}
    \item{\code{X2003031}}{a numeric vector}
    \item{\code{X2003032}}{a numeric vector}
    \item{\code{X2003040}}{a numeric vector}
    \item{\code{X2003041}}{a numeric vector}
    \item{\code{X2003042}}{a numeric vector}
    \item{\code{X2003050}}{a numeric vector}
    \item{\code{X2003051}}{a numeric vector}
    \item{\code{X2003052}}{a numeric vector}
    \item{\code{X2003060}}{a numeric vector}
    \item{\code{X2003061}}{a numeric vector}
    \item{\code{X2003062}}{a numeric vector}
    \item{\code{X2003070}}{a numeric vector}
    \item{\code{X2003071}}{a numeric vector}
    \item{\code{X2003072}}{a numeric vector}
    \item{\code{X2003080}}{a numeric vector}
    \item{\code{X2003081}}{a numeric vector}
    \item{\code{X2003082}}{a numeric vector}
    \item{\code{X2003090}}{a numeric vector}
    \item{\code{X2003091}}{a numeric vector}
    \item{\code{X2003092}}{a numeric vector}
    \item{\code{X2003100}}{a numeric vector}
    \item{\code{X2003101}}{a numeric vector}
    \item{\code{X2003102}}{a numeric vector}
    \item{\code{X2003110}}{a numeric vector}
    \item{\code{X2003111}}{a numeric vector}
    \item{\code{X2003112}}{a numeric vector}
    \item{\code{X2003120}}{a numeric vector}
    \item{\code{X2003121}}{a numeric vector}
    \item{\code{X2003122}}{a numeric vector}
    \item{\code{X2004010}}{a numeric vector}
    \item{\code{X2004011}}{a numeric vector}
    \item{\code{X2004012}}{a numeric vector}
    \item{\code{X2004020}}{a numeric vector}
    \item{\code{X2004021}}{a numeric vector}
    \item{\code{X2004022}}{a numeric vector}
    \item{\code{X2004030}}{a numeric vector}
    \item{\code{X2004031}}{a numeric vector}
    \item{\code{X2004032}}{a numeric vector}
    \item{\code{X2004040}}{a numeric vector}
    \item{\code{X2004041}}{a numeric vector}
    \item{\code{X2004042}}{a numeric vector}
    \item{\code{X2004050}}{a numeric vector}
    \item{\code{X2004051}}{a numeric vector}
    \item{\code{X2004052}}{a numeric vector}
    \item{\code{X2004060}}{a numeric vector}
    \item{\code{X2004061}}{a numeric vector}
    \item{\code{X2004062}}{a numeric vector}
    \item{\code{X2004070}}{a numeric vector}
    \item{\code{X2004071}}{a numeric vector}
    \item{\code{X2004072}}{a numeric vector}
    \item{\code{X2004080}}{a numeric vector}
    \item{\code{X2004081}}{a numeric vector}
    \item{\code{X2004082}}{a numeric vector}
    \item{\code{X2004090}}{a numeric vector}
    \item{\code{X2004091}}{a numeric vector}
    \item{\code{X2004092}}{a numeric vector}
    \item{\code{X2004100}}{a numeric vector}
    \item{\code{X2004101}}{a numeric vector}
    \item{\code{X2004102}}{a numeric vector}
    \item{\code{X2004110}}{a numeric vector}
    \item{\code{X2004111}}{a numeric vector}
    \item{\code{X2004112}}{a numeric vector}
    \item{\code{X2004120}}{a numeric vector}
    \item{\code{X2004121}}{a numeric vector}
    \item{\code{X2004122}}{a numeric vector}
    \item{\code{X2005010}}{a numeric vector}
    \item{\code{X2005011}}{a numeric vector}
    \item{\code{X2005012}}{a numeric vector}
    \item{\code{X2005020}}{a numeric vector}
    \item{\code{X2005021}}{a numeric vector}
    \item{\code{X2005022}}{a numeric vector}
    \item{\code{X2005030}}{a numeric vector}
    \item{\code{X2005031}}{a numeric vector}
    \item{\code{X2005032}}{a numeric vector}
    \item{\code{X2005040}}{a numeric vector}
    \item{\code{X2005041}}{a numeric vector}
    \item{\code{X2005042}}{a numeric vector}
    \item{\code{X2005050}}{a numeric vector}
    \item{\code{X2005051}}{a numeric vector}
    \item{\code{X2005052}}{a numeric vector}
    \item{\code{X2005060}}{a numeric vector}
    \item{\code{X2005061}}{a numeric vector}
    \item{\code{X2005062}}{a numeric vector}
    \item{\code{X2005070}}{a numeric vector}
    \item{\code{X2005071}}{a numeric vector}
    \item{\code{X2005072}}{a numeric vector}
    \item{\code{X2005080}}{a numeric vector}
    \item{\code{X2005081}}{a numeric vector}
    \item{\code{X2005082}}{a numeric vector}
    \item{\code{X2005090}}{a numeric vector}
    \item{\code{X2005091}}{a numeric vector}
    \item{\code{X2005092}}{a numeric vector}
    \item{\code{X2005100}}{a numeric vector}
    \item{\code{X2005101}}{a numeric vector}
    \item{\code{X2005102}}{a numeric vector}
    \item{\code{X2005110}}{a numeric vector}
    \item{\code{X2005111}}{a numeric vector}
    \item{\code{X2005112}}{a numeric vector}
    \item{\code{X2005120}}{a numeric vector}
    \item{\code{X2005121}}{a numeric vector}
    \item{\code{X2005122}}{a numeric vector}
    \item{\code{X2006010}}{a numeric vector}
    \item{\code{X2006011}}{a numeric vector}
    \item{\code{X2006012}}{a numeric vector}
    \item{\code{X2006020}}{a numeric vector}
    \item{\code{X2006021}}{a numeric vector}
    \item{\code{X2006022}}{a numeric vector}
    \item{\code{X2006030}}{a numeric vector}
    \item{\code{X2006031}}{a numeric vector}
    \item{\code{X2006032}}{a numeric vector}
    \item{\code{X2006040}}{a numeric vector}
    \item{\code{X2006041}}{a numeric vector}
    \item{\code{X2006042}}{a numeric vector}
    \item{\code{X2006050}}{a numeric vector}
    \item{\code{X2006051}}{a numeric vector}
    \item{\code{X2006052}}{a numeric vector}
    \item{\code{X2006060}}{a numeric vector}
    \item{\code{X2006061}}{a numeric vector}
    \item{\code{X2006062}}{a numeric vector}
    \item{\code{X2006070}}{a numeric vector}
    \item{\code{X2006071}}{a numeric vector}
    \item{\code{X2006072}}{a numeric vector}
    \item{\code{X2006080}}{a numeric vector}
    \item{\code{X2006081}}{a numeric vector}
    \item{\code{X2006082}}{a numeric vector}
    \item{\code{X2006090}}{a numeric vector}
    \item{\code{X2006091}}{a numeric vector}
    \item{\code{X2006092}}{a numeric vector}
    \item{\code{X2006100}}{a numeric vector}
    \item{\code{X2006101}}{a numeric vector}
    \item{\code{X2006102}}{a numeric vector}
    \item{\code{X2006110}}{a numeric vector}
    \item{\code{X2006111}}{a numeric vector}
    \item{\code{X2006112}}{a numeric vector}
    \item{\code{X2006120}}{a numeric vector}
    \item{\code{X2006121}}{a numeric vector}
    \item{\code{X2006122}}{a numeric vector}
    \item{\code{X2007010}}{a numeric vector}
    \item{\code{X2007011}}{a numeric vector}
    \item{\code{X2007012}}{a numeric vector}
    \item{\code{X2007020}}{a numeric vector}
    \item{\code{X2007021}}{a numeric vector}
    \item{\code{X2007022}}{a numeric vector}
    \item{\code{X2007030}}{a numeric vector}
    \item{\code{X2007031}}{a numeric vector}
    \item{\code{X2007032}}{a numeric vector}
    \item{\code{X2007040}}{a numeric vector}
    \item{\code{X2007041}}{a numeric vector}
    \item{\code{X2007042}}{a numeric vector}
    \item{\code{X2007050}}{a numeric vector}
    \item{\code{X2007051}}{a numeric vector}
    \item{\code{X2007052}}{a numeric vector}
    \item{\code{X2007060}}{a numeric vector}
    \item{\code{X2007061}}{a numeric vector}
    \item{\code{X2007062}}{a numeric vector}
    \item{\code{X2007070}}{a numeric vector}
    \item{\code{X2007071}}{a numeric vector}
    \item{\code{X2007072}}{a numeric vector}
    \item{\code{X2007080}}{a numeric vector}
    \item{\code{X2007081}}{a numeric vector}
    \item{\code{X2007082}}{a numeric vector}
    \item{\code{X2007090}}{a numeric vector}
    \item{\code{X2007091}}{a numeric vector}
    \item{\code{X2007092}}{a numeric vector}
    \item{\code{X2007100}}{a numeric vector}
    \item{\code{X2007101}}{a numeric vector}
    \item{\code{X2007102}}{a numeric vector}
    \item{\code{X2007110}}{a numeric vector}
    \item{\code{X2007111}}{a numeric vector}
    \item{\code{X2007112}}{a numeric vector}
    \item{\code{X2007120}}{a numeric vector}
    \item{\code{X2007121}}{a numeric vector}
    \item{\code{X2007122}}{a numeric vector}
    \item{\code{X2008010}}{a numeric vector}
    \item{\code{X2008011}}{a numeric vector}
    \item{\code{X2008012}}{a numeric vector}
    \item{\code{X2008020}}{a numeric vector}
    \item{\code{X2008021}}{a numeric vector}
    \item{\code{X2008022}}{a numeric vector}
    \item{\code{X2008030}}{a numeric vector}
    \item{\code{X2008031}}{a numeric vector}
    \item{\code{X2008032}}{a numeric vector}
    \item{\code{X2008040}}{a numeric vector}
    \item{\code{X2008041}}{a numeric vector}
    \item{\code{X2008042}}{a numeric vector}
    \item{\code{X2008050}}{a numeric vector}
    \item{\code{X2008051}}{a numeric vector}
    \item{\code{X2008052}}{a numeric vector}
    \item{\code{X2008060}}{a numeric vector}
    \item{\code{X2008061}}{a numeric vector}
    \item{\code{X2008062}}{a numeric vector}
    \item{\code{X2008070}}{a numeric vector}
    \item{\code{X2008071}}{a numeric vector}
    \item{\code{X2008072}}{a numeric vector}
    \item{\code{X2008080}}{a numeric vector}
    \item{\code{X2008081}}{a numeric vector}
    \item{\code{X2008082}}{a numeric vector}
    \item{\code{X2008090}}{a numeric vector}
    \item{\code{X2008091}}{a numeric vector}
    \item{\code{X2008092}}{a numeric vector}
    \item{\code{X2008100}}{a numeric vector}
    \item{\code{X2008101}}{a numeric vector}
    \item{\code{X2008102}}{a numeric vector}
    \item{\code{X2008110}}{a numeric vector}
    \item{\code{X2008111}}{a numeric vector}
    \item{\code{X2008112}}{a numeric vector}
    \item{\code{X2008120}}{a numeric vector}
    \item{\code{X2008121}}{a numeric vector}
    \item{\code{X2008122}}{a numeric vector}
    \item{\code{X2009010}}{a numeric vector}
    \item{\code{X2009011}}{a numeric vector}
    \item{\code{X2009012}}{a numeric vector}
    \item{\code{X2009020}}{a numeric vector}
    \item{\code{X2009021}}{a numeric vector}
    \item{\code{X2009022}}{a numeric vector}
    \item{\code{X2009030}}{a numeric vector}
    \item{\code{X2009031}}{a numeric vector}
    \item{\code{X2009032}}{a numeric vector}
    \item{\code{X2009040}}{a numeric vector}
    \item{\code{X2009041}}{a numeric vector}
    \item{\code{X2009042}}{a numeric vector}
    \item{\code{X2009050}}{a numeric vector}
    \item{\code{X2009051}}{a numeric vector}
    \item{\code{X2009052}}{a numeric vector}
    \item{\code{X2009060}}{a numeric vector}
    \item{\code{X2009061}}{a numeric vector}
    \item{\code{X2009062}}{a numeric vector}
    \item{\code{X2009070}}{a numeric vector}
    \item{\code{X2009071}}{a numeric vector}
    \item{\code{X2009072}}{a numeric vector}
    \item{\code{X2009080}}{a numeric vector}
    \item{\code{X2009081}}{a numeric vector}
    \item{\code{X2009082}}{a numeric vector}
    \item{\code{X2009090}}{a numeric vector}
    \item{\code{X2009091}}{a numeric vector}
    \item{\code{X2009092}}{a numeric vector}
    \item{\code{X2009100}}{a numeric vector}
    \item{\code{X2009101}}{a numeric vector}
    \item{\code{X2009102}}{a numeric vector}
    \item{\code{X2009110}}{a numeric vector}
    \item{\code{X2009111}}{a numeric vector}
    \item{\code{X2009112}}{a numeric vector}
    \item{\code{X2009120}}{a numeric vector}
    \item{\code{X2009121}}{a numeric vector}
    \item{\code{X2009122}}{a numeric vector}
  }
}
\details{
no details yet.
}
\source{
VITO
}
\examples{
data(timeseries)
## maybe str(timeseries) ; plot(timeseries) ...
}
\keyword{datasets}
